triage-agent/requirements.txt
flask==3.0.0
google-cloud-pubsub==2.21.0
google-cloud-aiplatform==1.50.0
google-cloud-firestore==2.16.0
